+++ backend/app/services/auth/authorization.js
@@ -15,3 +15,19 @@ export function domainAuthorization(allowedDomains) {
     return allowedDomainsSet.has(domain);
   };
 }
+
+export function userAuthorization(allowedUsers) {
+  const allowedUsersSet = new Set(allowedUsers);
+
+  return ({ email }) => {
+    if (allowedUsersSet.size === 0) {
+      return true;
+    }
+
+    if (email === undefined) {
+      return false
+    }
+
+    return allowedUsersSet.has(email);
+  };
+}